feat: it will now compile an .exe file that does nothing. This is almost certainly due to my best-guess attempts at making winapi stuff work based on ancient and fragmented msdn posts from ~2005-2012. Also I couldnt get transworkshop working in 64bit despite adding steam_api64.lib, but its not needed to compile transcendence.exe so I left it broken in this commit.

+++ Transcendence/TransData/Words.cpp
@@ -164,7 +164,7 @@ void AddWord (TraverseCtx &Ctx, const CString &sWord)
 
 	else
 		{
-		Ctx.pWordList->AddEntry(strToLower(sWord), NULL);
+		Ctx.pWordList->AddEntry(strToLower(sWord), (size_t)NULL);
 		}
 	}
